requirements.txt
apache-beam==2.43.0
black==22.10.0
cachetools==4.2.4
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.2.0
crcmod==1.7
dill==0.3.1.1
docopt==0.6.2
fastavro==1.7.0
fasteners==0.18
google-api-core==2.8.2
google-apitools==0.5.31
google-auth==2.15.0
google-auth-httplib2==0.1.0
google-cloud-bigquery==3.3.3
google-cloud-bigquery-storage==2.13.2
google-cloud-bigtable==1.7.3
google-cloud-core==2.3.2
google-cloud-datastore==1.15.5
google-cloud-dlp==3.9.0
google-cloud-language==1.3.2
google-cloud-pubsub==2.13.7
google-cloud-pubsublite==1.6.0
google-cloud-recommendations-ai==0.7.1
google-cloud-spanner==3.22.0
google-cloud-videointelligence==1.16.3
google-cloud-vision==1.0.2
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.56.4
grpc-google-iam-v1==0.12.4
grpcio==1.51.1
grpcio-status==1.48.2
hdfs==2.7.0
httplib2==0.20.4
idna==3.4
mypy-extensions==0.4.3
numpy==1.22.4
oauth2client==4.1.3
objsize==0.5.2
orjson==3.8.3
overrides==6.5.0
packaging==21.3
pandas==1.5.2
pathspec==0.10.2
platformdirs==2.5.4
proto-plus==1.22.1
protobuf==3.19.4
pyarrow==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydot==1.4.2
pymongo==3.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
regex==2022.10.31
requests==2.28.1
rsa==4.9
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.13
zstandard==0.19.0
